+++ superset/assets/src/SqlLab/components/SouthPane.jsx
@@ -20,6 +20,7 @@ const propTypes = {
   actions: PropTypes.object.isRequired,
   activeSouthPaneTab: PropTypes.string,
   height: PropTypes.number,
+  databases: PropTypes.object.isRequired,
 };
 
 const defaultProps = {
@@ -46,6 +47,7 @@ class SouthPane extends React.PureComponent {
           query={latestQuery}
           actions={props.actions}
           height={innerTabHeight}
+          database={this.props.databases[latestQuery.dbId]}
         />
       );
     } else {
@@ -100,6 +102,7 @@ class SouthPane extends React.PureComponent {
 function mapStateToProps({ sqlLab }) {
   return {
     activeSouthPaneTab: sqlLab.activeSouthPaneTab,
+    databases: sqlLab.databases,
   };
 }
