+++ superset/assets/src/SqlLab/constants.js
@@ -1,5 +1,3 @@
-import { t } from '../locales';
-
 export const STATE_BSSTYLE_MAP = {
   failed: 'danger',
   pending: 'info',
@@ -25,10 +23,3 @@ export const TIME_OPTIONS = [
   '90 days ago',
   '1 year ago',
 ];
-
-export const VISUALIZE_VALIDATION_ERRORS = {
-  REQUIRE_CHART_TYPE: t('Pick a chart type!'),
-  REQUIRE_TIME: t('To use this chart type you need at least one column flagged as a date'),
-  REQUIRE_DIMENSION: t('To use this chart type you need at least one dimension'),
-  REQUIRE_AGGREGATION_FUNCTION: t('To use this chart type you need at least one aggregation function'),
-};
