+++ superset/db_engine_specs.py
@@ -100,6 +100,7 @@ class BaseEngineSpec(object):
     limit_method = LimitMethod.FORCE_LIMIT
     time_secondary_columns = False
     inner_joins = True
+    allows_subquery = True
 
     @classmethod
     def get_time_grains(cls):
@@ -1368,6 +1369,7 @@ class DruidEngineSpec(BaseEngineSpec):
     """Engine spec for Druid.io"""
     engine = 'druid'
     inner_joins = False
+    allows_subquery = False
 
     time_grain_functions = {
         None: '{col}',
